+++ src/public/components/content.html
@@ -1,3 +1,3 @@
 <section>
-    
+    content
 </section>
