+++ src/public/components/leftNavBar.html
@@ -1,3 +1,3 @@
 <nav>
-    
+    leftnavbar
 </nav>
