+++ src/public/components/nav.html
@@ -1,5 +1,5 @@
 <nav>
     <header>
-        
+        nav
     </header>
 </nav>
