+++ src/public/components/rightNavBar.html
@@ -1,3 +1,3 @@
 <nav>
-    
+    rightnavbar
 </nav>
