+++ src/public/index.html
@@ -6,7 +6,7 @@
     <meta name="viewport" content="width=device-width, initial-scale=1.0">
     <title>Ingenieria I y Grupo 3</title>
     <link rel="stylesheet" href="./assets/css/main.css" />
-    <script src="./js/main.js" type="module"></script>
+    <script type="module" src="./assets/js/main.js"></script> 
 </head>
 
 <body id="root">
